epjitsu: x,y_resolution difference treatment
ScanScanp has a resolution 225x200 in each model.

The function which treat it should use a mathematic theory.
Euclidean algorithm should be used to get GCD (Greatest Common Devisor)
for this calcuration.

This patch make a separate function that should be get_GCD and
simplize a logic in copy image function.

TBD: this don't have Euclidean algorithm. Just prepare for it.

Signed-off-by: Hiroshi Miura <miurahr@linux.com>